feat(io): Change OptrionalRequestData To entity in request
