+++ src/view/cli/input/tests_parser.rs
@@ -2,8 +2,8 @@ use std::collections::HashMap;
 
 use super::clap_definition::root_command;
 use super::parser::parse_clap_input_to_commands;
-use crate::app::services::request::entity::{RequestData, METHODS};
-use crate::view::cli::commands::{CliCommand, OptionalRequestData};
+use crate::app::services::request::entity::{OptionalRequestData, RequestData, METHODS};
+use crate::view::cli::commands::CliCommand;
 
 #[test]
 fn test_parse_all_methods_command() {
